docs: Refactor Foundation Pages (TabContent -> MainContent) (MainContent -> TopContent), Seperation Top and Content
