+++ packages/docs/src/app/_components/layout/MainContent/MainContent.tsx
@@ -1,45 +1,44 @@
 import type { FC } from "react";
 import React from "react";
-import Heading from "@mittwald/flow-react-components/Heading";
-import { Link } from "@mittwald/flow-react-components/Link";
-import { LayoutCard } from "@mittwald/flow-react-components/LayoutCard";
-import { Section } from "@mittwald/flow-react-components/Section";
-import { ColumnLayout } from "@mittwald/flow-react-components/ColumnLayout";
-import { IconExternalLink } from "@mittwald/flow-react-components/Icons";
 import type { MdxFile } from "@/lib/mdx/MdxFile";
 import MdxFileView from "@/lib/mdx/components/MdxFileView";
+import { Tab, TabTitle } from "@mittwald/flow-react-components/Tabs";
+import styles from "../../../layout.module.scss";
+import { MdxFileFactory } from "@/lib/mdx/MdxFileFactory";
+import * as path from "path";
 
 interface Props {
   mdxFile: MdxFile;
+  tabTitle?: string;
+  showTabs?: boolean;
 }
 
-export const MainContent: FC<Props> = (props) => {
-  const { mdxFile } = props;
+export const MainContent: FC<Props> = async (props) => {
+  const { mdxFile, tabTitle, showTabs } = props;
 
-  const component = mdxFile.mdxSource.frontmatter.component;
-
-  return (
-    <>
-      <LayoutCard>
-        <ColumnLayout l={component ? [1, 1] : [1]} m={[1]}>
-          <Section>
-            <Heading level={1}>{mdxFile.getTitle()}</Heading>
-
-            {mdxFile.mdxSource.frontmatter.description}
-
-            {component && (
-              <Link href={mdxFile.getGitHubUrl()}>
-                GitHub
-                <IconExternalLink />
-              </Link>
-            )}
-          </Section>
-
-          <MdxFileView mdxFile={mdxFile.serialize()} />
-        </ColumnLayout>
-      </LayoutCard>
-    </>
+  const indexFile = await MdxFileFactory.fromFile(
+    "src/content/" + path.dirname(mdxFile.filename) + "/index.mdx",
   );
+
+  const content = (
+    <div className={styles.tabContent}>
+      <MdxFileView
+        mdxFile={mdxFile.serialize()}
+        indexFile={indexFile.serialize()}
+      />
+    </div>
+  );
+
+  if (showTabs) {
+    return (
+      <Tab>
+        <TabTitle>{tabTitle}</TabTitle>
+        {content}
+      </Tab>
+    );
+  }
+
+  return <> {content} </>;
 };
 
 export default MainContent;
