--- packages/docs/src/app/_components/layout/TabContent/index.ts
@@ -1,3 +1,0 @@
-import TabContent from "@/app/_components/layout/TabContent/TabContent";
-
-export default TabContent;
