+++ packages/docs/src/app/_components/layout/TopContent/index.ts
@@ -1,0 +1,3 @@
+import TopContent from "@/app/_components/layout/TopContent/TopContent";
+
+export default TopContent;
