+++ packages/docs/src/content/02-foundations/content-guidelines/sprach-guide/overview.mdx
@@ -1,0 +1,162 @@
+## Wie wir sprechen
+
+### Technisch aber einfach verständlich
+
+Wir überzeugen mit Inhalten, nicht mit langen und verschachtelten Erklärungen.
+Kurze Sätze, einfacher Satzbau – so bleiben wir jederzeit übersichtlich und
+verständlich.
+
+### Auf Augenhöhe
+
+Wir drücken uns technisch aber verständlich aus – genau mit der technischen
+Tiefe, die dem Nutzer weiterhilft.
+
+### Aktiv
+
+Wir machen Technologie zugänglich. Das bietet keinen Raum für Passivität. Dafür
+kommunizieren wir aktiv und suchen proaktiv nach den besten Lösungen. So
+generieren wir das beste Nutzungserlebnis.
+
+### Konsistent
+
+Kunde oder Nutzer? Buchen oder installieren? Ist alles geregelt. Deshalb quäle
+dich nicht mit solchen Fragen, sondern verwende die allgemeinen Schreibweisen
+des Unternehmens. So generieren wir für unsere Nutzer immer und überall das
+gleiche Erlebnis.
+
+### Glossar
+
+Sie können sich darauf verlassen, dass wir ihnen in jeder Situation weiterhelfen
+– mit einer Sprache, die sie von uns kennen und erwarten.
+
+---
+
+## Leitlinien
+
+### Aktive Sprache
+
+Wir agieren aktiv. Das zeigen wir auch in unserer Sprache. Deshalb verzichten
+wir auf passive Satzkonstruktionen.
+
+<DoAndDont>
+  <Do>Paula hat den Grill angeschmissen.</Do>
+  <Dont>Der Grill wurde von Paula angeschmissen.</Dont>
+</DoAndDont>
+
+### Sie, Du oder Ihr
+
+Respekt braucht kein „Sie“. Wir reden Nutzer genauso an, wie Kolleg\*innen: mit
+einem freundlichen „Du“. Unsere Nutzer reden wir persönlich an, also im
+Singular.
+
+In Mailings und Tickets werden Du, Dein, Deiner, usw. immer großgeschrieben.
+
+<DoAndDont>
+  <Do>Du weißt, mittwald hat das beste Hosting für Agenturen.</Do>
+  <Dont>Ihr wisst, mittwald hat das beste Hosting für Agenturen.</Dont>
+</DoAndDont>
+
+### Bindestrich: Wörter koppeln
+
+Zweisprachige Wortkombinationen werden mit einem Bindestrich gekoppelt.
+
+Englische Begriffe werden nicht gekoppelt, deutsche Kombinationen werden
+zusammengeschrieben.
+
+<DoAndDont>
+  <Do>Ein Tarif-Upgrade kommt selten allein. Erst recht im Cloud Hosting.</Do>
+  <Dont>
+    Ein Tarif Upgrade allein macht noch keinen Sommer. Auch nicht im
+    Cloud-Hosting.
+  </Dont>
+</DoAndDont>
+
+Egal ob mittwald Hosting oder mStudio Login – die Regel ist immer die gleiche:
+mittwald und Produktnamen werden niemals mit einem Bindestrich an ein anderes
+Wort gekoppelt.
+
+Generell gilt: Im Zweifel für die Lesbarkeit. Macht die korrekte Schreibweise
+Probleme (sieht schlecht aus, Umbruch, usw.), missachten wir die Regeln. Das
+sollte aber nur die Ausnahme sein.
+
+### Gendern
+
+mittwald lebt Gleichberechtigung. Um kurz und prägnant formulieren zu können,
+gendern wir jedoch nur dort, wo es Sinn macht, z.B. in Stellenanzeigen.
+
+An Orten mit wenig Platz, etwa im mStudio oder auf Social Media, verwenden wir
+die meist verwendete Schreibweise.
+
+### KEIN! GRUND! ZU! SCHREIEN!!!
+
+Was wir tun ist spannend genug. Da braucht es keine Versalien und
+Ausrufezeichen. Wie immer gilt auch hier: Ausnahmen bestätigen die Regel. Danke
+fürs Lesen!
+
+## Klare Gliederung
+
+Dein Text sollte durch Absätze gegliedert sein. Im besten Fall sind diese nicht
+länger als 5 Zeilen. Auch Nummerierungen oder Bulletpoints helfen den Lesern,
+deinem Text besser zu folgen und Wichtiges schneller wahrzunehmen.
+
+### Kurze Sätze
+
+Fass dich kurz. Um Technisches einfach zu formulieren, helfen kurze Sätze.
+Faustregel: Satzlänge maximal 1,5 Zeilen. 1 Subjekt pro Satz.
+
+### mittwald, mStudio und proSpace werden IMMER klein geschrieben
+
+... auch am Satzanfang.
+
+### Wichtige Aspekte hervorheben
+
+Fette zentrale Begriffe, um Wichtiges in den Vordergrund zu stellen. Beschränke
+dich dabei auf einzelne Wörter. Ganze Sätze (oder gar Absätze) werden nicht
+gefettet.
+
+### Zahlen
+
+Um die Lesbarkeit zu steigern, schreiben wir alle Zahlen als Ziffern, auch von 0
+bis 12.
+
+---
+
+## Glossar
+
+- **Anlegen:** Hinzufügen von Objekten ohne Vertragsposition, die von Grund auf
+  von den Nutzern erstellt werden müssen.
+- **Artikel:** Alle kostenpflichtigen Produkte, z.B. Domains, die hinzugebucht
+  werden.
+- **Buchen:** Beschreibt das kostenpflichtige Bestellen eines Artikels,
+  Projektes oder Servers. Wird auch als hinzufügen / anlegen / kaufen /
+  Kaufprozess starten / usw. beschrieben, jeweils ergänzt durch das zu buchende
+  Objekt (Beispiel: Domain buchen).
+- **Installieren:** Siehe anlegen, nur speziell für den Bereich Software (Apps,
+  Systemsoftware).
+- **Kündigen:** Elemente, die an einer vertraglichen Verbindlichkeit hängen,
+  müssen gekündigt werden (Ausnahme: Domains).
+- **Löschen:** Elemente, die nicht an einer vertraglichen Verbindlichkeit
+  hängen, werden gelöscht (Ausnahme: Domains).
+- **Nutzer:** Wir reden im mStudio immer vom Nutzer oder der Organisation. Den
+  Begriff "Kunde" verwenden wir nicht.
+- **Organisation:** In der Organisation werden alle wichtigen
+  Zahlungsinformationen sowie Vertragspartnerdaten hinterlegt. Server und
+  Einzelprojekte sind immer einer Organisation zugeordnet.
+- **Projekt:** Das Projekt beinhaltet alle Komponenten zum Betreiben eines
+  Webprojektes. Bspw. Apps, Domains, Datenbanken, E-Mail-Postfächer etc.
+- **Ressourcen:** Beschreibt die Kombination aus vCPU/RAM (Flavour) und
+  Speicherplatz.
+- **Server:** Ein Server ist einer Organisation zugeordnet und dient zur
+  Verwaltung des Speicherplatzes und der Ressourcen. Auf einem Server können,
+  solange die technischen Limits nicht überschritten werden, unbegrenzt viele
+  Projekte angelegt werden.
+- **Ticket:** Schriftliche Anfrage eines Nutzers.
+- **Übersicht:** Eine Liste beliebiger Objekte (z.B. Artikel, Domains, Projekte,
+  E-Mail-Adressen, Apps, …).
+- **USt.:** Alle Preise, die dem Nutzer in der finalen Buchungsübersicht
+  angezeigt werden, werden in Netto und Brutto dargestellt. Dabei werden die
+  Bruttowerte mit dem Zusatz “inkl. USt.” dargestellt.
+- **Vertragspartner:** Eine natürliche oder juristische Person, die mit uns
+  einen Vertrag abgeschlossen hat (es gibt nicht zwangsläufig einen Nutzer, der
+  mit dieser Person übereinstimmt).
+- **Zahlungsart:** Auswahl der Optionen, wie Produkte bezahlt werden können.
