+++ packages/docs/src/content/02-foundations/design/typography/index.mdx
@@ -1,74 +1,1 @@
-# Fonts
-
-Für Benutzeroberflächen wird hauptsächlich die Font **Inter** verwendet. Sie
-wird z. B. in [Headings](/03-components/content/heading), im Body und in den
-Components eingesetzt. Im Code wird die Font Size immer in REM angegeben, damit
-die Schrift vom User leicht vergrößert werden kann. <br/> Die Line Height
-beträgt immer min. 150 % der Höhe der Font.
-
----
-
-# Headings
-
-[Headings](/03-components/content/heading) zeichnen sich durch ihre Farbe und
-die große Auswahl an verschiedenen Größen zur Strukturierung aus.
-
-|                                        | REM       | PX    | Line Height |
-| -------------------------------------- | --------- | ----- | ----------- |
-| <Heading level={1}>Heading 1</Heading> | 2 rem     | 32 px | 48 lh       |
-| <Heading level={2}>Heading 2</Heading> | 1,25 rem  | 20 px | 30 lh       |
-| <Heading level={3}>Heading 3</Heading> | 1.125 rem | 18 px | 27 lh       |
-| <Heading level={4}>Heading 4</Heading> | 1 rem     | 16 px | 24 lh       |
-| <Heading level={5}>Heading 5</Heading> | 1 rem     | 16 px | 24 lh       |
-
----
-
-# Text
-
-Text wird überall auf Benutzeroberflächen verwendet, vom Fließtext bis hin zu
-kleinen Hinweisen. Bei Small kann zwischen den Line Heights 21 und 24 gewählt
-werden, da gerade 24lh gut in das Layout passt und 21lh für kleinere Höhen
-besser geeignet ist.
-
-|                                 | REM       | PX    | Line Height   |
-| ------------------------------- | --------- | ----- | ------------- |
-| Text (Body)                     | 1 rem     | 16 px | 24 lh         |
-| **Text (Body)**                 | 1 rem     | 16 px | 24 lh         |
-| <small>Text (Small)</small>     | 0,875 rem | 14 px | 21 lh / 24 lh |
-| **<small>Text (Small)</small>** | 0,875 rem | 14 px | 21 lh / 24 lh |
-
----
-
-# Label
-
-[Labels](03-components/content/label) sind durch ihre graue Farbe und
-Bold-Darstellung als beschreibendes Element in komplexeren Components gut
-erkennbar.
-
-|                      | REM       | PX    | Line Height   |
-| -------------------- | --------- | ----- | ------------- |
-| <Label>Label</Label> | 0,875 rem | 14 px | 21 lh / 24 lh |
-
----
-
-# Links
-
-Frei stehende [Links](/03-components/navigation/link) werden in Bold
-dargestellt. Im Fließtext stehende Links verfügen über einen Unterstrich, um sie
-im Text besser erkennen zu können (s. Accessibility).
-
-|                   | REM       | PX    | Line Height |
-| ----------------- | --------- | ----- | ----------- |
-| <Link>Link</Link> | 0,875 rem | 14 px | 24 lh       |
-| [Link]()          | 1 rem     | 16 px | 24 lh       |
-
----
-
-# Code
-
-Um Code von der anderen Font abzuheben, wird bei Code **Fira Code** verwendet,
-da sie einen eher technisches Aussehen hat.
-
-|        | REM       | PX    | Line Height |
-| ------ | --------- | ----- | ----------- |
-| `code` | 0,875 rem | 14 px | 24 lh       |
+TODO Typography TEXT
